third_party/blink/web_tests/platform/linux/external/wpt/url/url-constructor.any.worker_exclude=(file_javascript_mailto)-expected.txt

This is a testharness.js-based test.
Found 156 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Parsing: <non-special://test:@test/x> without base
  assert_equals: href expected "non-special://test@test/x" but got "non-special://test:@test/x"
[FAIL] Parsing: <non-special://:@test/x> without base
  assert_equals: href expected "non-special://test/x" but got "non-special://:@test/x"
[FAIL] Parsing: <non-special://f:999999/c> against <http://example.org/foo/bar>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <foo://> against <http://example.org/foo/bar>
  assert_equals: pathname expected "" but got "//"
[FAIL] Parsing: <foo://///////> against <http://example.org/foo/bar>
  assert_equals: pathname expected "///////" but got "/////////"
[FAIL] Parsing: <foo://///////bar.com/> against <http://example.org/foo/bar>
  assert_equals: pathname expected "///////bar.com/" but got "/////////bar.com/"
[FAIL] Parsing: <foo:////://///> against <http://example.org/foo/bar>
  assert_equals: pathname expected "//://///" but got "////://///"
[FAIL] Parsing: <ssh://example.com/foo/bar.git> against <http://example.org/>
  assert_equals: host expected "example.com" but got ""
[FAIL] Parsing: <http://example.com/foo/%2e%2> without base
  assert_equals: href expected "http://example.com/foo/%2e%2" but got "http://example.com/foo/.%2"
[FAIL] Parsing: <http://example.com/foo/%2e./%2e%2e/.%2e/%2e.bar> without base
  assert_equals: href expected "http://example.com/%2e.bar" but got "http://example.com/..bar"
[FAIL] Parsing: <http://www/foo%2Ehtml> without base
  assert_equals: href expected "http://www/foo%2Ehtml" but got "http://www/foo.html"
[FAIL] Parsing: <httpa://foo:80/> without base
  assert_equals: host expected "foo:80" but got ""
[FAIL] Parsing: <gopher://foo:70/> without base
  assert_equals: host expected "foo:70" but got ""
[FAIL] Parsing: <gopher://foo:443/> without base
  assert_equals: host expected "foo:443" but got ""
[FAIL] Parsing: <http://example example.com> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://GOO 　goo.com> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://a.b.c.xn--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://10.0.0.xn--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://a.b.c.XN--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://a.b.c.Xn--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://10.0.0.XN--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://10.0.0.xN--pokxncvks> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://％４１.com> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://%ef%bc%85%ef%bc%94%ef%bc%91.com> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://faß.ExAmPlE/> without base
  assert_equals: host expected "fa%C3%9F.ExAmPlE" but got ""
[FAIL] Parsing: <http://192.168.0.1 hello> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <https://x x:12> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <h://.> without base
  assert_equals: host expected "." but got ""
[FAIL] Parsing: <http://[::%31]> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://%5B::1]> against <http://other.com/>
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <notspecial://host/?'> without base
  assert_equals: href expected "notspecial://host/?'" but got "notspecial://host/?%27"
[FAIL] Parsing: <i> against <sc://ho/pa>
  assert_equals: host expected "ho" but got ""
[FAIL] Parsing: <i> against <sc:///pa/pa>
  assert_equals: pathname expected "/pa/i" but got "///pa/i"
[FAIL] Parsing: <../i> against <sc://ho/pa>
  assert_equals: host expected "ho" but got ""
[FAIL] Parsing: <../i> against <sc:///pa/pa>
  assert_equals: href expected "sc:///i" but got "sc:///pa/i"
[FAIL] Parsing: </i> against <sc://ho/pa>
  assert_equals: host expected "ho" but got ""
[FAIL] Parsing: </i> against <sc:///pa/pa>
  assert_equals: href expected "sc:///i" but got "sc:///pa/i"
[FAIL] Parsing: <?i> against <sc://ho/pa>
  assert_equals: host expected "ho" but got ""
[FAIL] Parsing: <?i> against <sc:///pa/pa>
  assert_equals: pathname expected "/pa/pa" but got "///pa/pa"
[FAIL] Parsing: <#i> against <sc://ho/pa>
  assert_equals: host expected "ho" but got ""
[FAIL] Parsing: <#i> against <sc:///pa/pa>
  assert_equals: pathname expected "/pa/pa" but got "///pa/pa"
[FAIL] Parsing: <about:/../> without base
  assert_equals: href expected "about:/" but got "about:/../"
[FAIL] Parsing: <data:/../> without base
  assert_equals: href expected "data:/" but got "data:/../"
[FAIL] Parsing: <sc://ñ.test/> without base
  assert_equals: host expected "%C3%B1.test" but got ""
[FAIL] Parsing: <sc://%/> without base
  assert_equals: host expected "%" but got ""
[FAIL] Parsing: <sc://@/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://te@s:t@/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://:/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://:12/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <x> against <sc://ñ>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <sc://a\0b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a<b> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a>b> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a[b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a\\b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a]b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a^b> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <sc://a|b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <foo://ho	st/> without base
  assert_equals: host expected "host" but got ""
[FAIL] Parsing: <foo://ho\nst/> without base
  assert_equals: host expected "host" but got ""
[FAIL] Parsing: <foo://ho\rst/> without base
  assert_equals: host expected "host" but got ""
[FAIL] Parsing: <http://a b/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://ho%20st/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <http://!"$&'()*+,-.;=_`{}~/> without base
  assert_equals: href expected "http://!\\"$&'()*+,-.;=_`{}~/" but got "http://!\\"$&'()%2A+,-.;=_`{}~/"
[FAIL] Parsing: <sc://!"$%&'()*+,-.;=_`{}~/> without base
  assert_equals: host expected "%01%02%03%04%05%06%07%08%0B%0C%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%7F!\\"$%&'()*+,-.;=_`{}~" but got ""
[FAIL] Parsing: <asdf://%43|/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <asdf://%43%7C/> without base
  assert_equals: host expected "%43%7C" but got ""
[FAIL] Parsing: <sc://ñ> without base
  assert_equals: host expected "%C3%B1" but got ""
[FAIL] Parsing: <sc://ñ?x> without base
  assert_equals: host expected "%C3%B1" but got ""
[FAIL] Parsing: <sc://ñ#x> without base
  assert_equals: host expected "%C3%B1" but got ""
[FAIL] Parsing: <#x> against <sc://ñ>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <?x> against <sc://ñ>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <sc://?> without base
  assert_equals: pathname expected "" but got "//"
[FAIL] Parsing: <sc://#> without base
  assert_equals: pathname expected "" but got "//"
[FAIL] Parsing: <///> against <sc://x/>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <////> against <sc://x/>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <////x/> against <sc://x/>
  assert_equals: href expected "sc:////x/" but got "sc://x/"
[FAIL] Parsing: <tftp://foobar.com/someconfig;mode=netascii> without base
  assert_equals: host expected "foobar.com" but got ""
[FAIL] Parsing: <telnet://user:pass@foobar.com:23/> without base
  assert_equals: username expected "user" but got ""
[FAIL] Parsing: <ut2004://10.10.10.10:7777/Index.ut2> without base
  assert_equals: host expected "10.10.10.10:7777" but got ""
[FAIL] Parsing: <redis://foo:bar@somehost:6379/0?baz=bam&qux=baz> without base
  assert_equals: username expected "foo" but got ""
[FAIL] Parsing: <rsync://foo@host:911/sup> without base
  assert_equals: username expected "foo" but got ""
[FAIL] Parsing: <git://github.com/foo/bar.git> without base
  assert_equals: host expected "github.com" but got ""
[FAIL] Parsing: <irc://myserver.com:6999/channel?passwd> without base
  assert_equals: host expected "myserver.com:6999" but got ""
[FAIL] Parsing: <dns://fw.example.org:9999/foo.bar.org?type=TXT> without base
  assert_equals: host expected "fw.example.org:9999" but got ""
[FAIL] Parsing: <ldap://localhost:389/ou=People,o=JNDITutorial> without base
  assert_equals: host expected "localhost:389" but got ""
[FAIL] Parsing: <git+https://github.com/foo/bar> without base
  assert_equals: host expected "github.com" but got ""
[FAIL] Parsing: <non-spec:/.//> without base
  assert_equals: pathname expected "//" but got "/.//"
[FAIL] Parsing: <non-spec:/..//> without base
  assert_equals: href expected "non-spec:/.//" but got "non-spec:/..//"
[FAIL] Parsing: <non-spec:/a/..//> without base
  assert_equals: href expected "non-spec:/.//" but got "non-spec:/a/..//"
[FAIL] Parsing: <non-spec:/.//path> without base
  assert_equals: pathname expected "//path" but got "/.//path"
[FAIL] Parsing: <non-spec:/..//path> without base
  assert_equals: href expected "non-spec:/.//path" but got "non-spec:/..//path"
[FAIL] Parsing: <non-spec:/a/..//path> without base
  assert_equals: href expected "non-spec:/.//path" but got "non-spec:/a/..//path"
[FAIL] Parsing: </.//path> against <non-spec:/p>
  assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
[FAIL] Parsing: </..//path> against <non-spec:/p>
  assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
[FAIL] Parsing: <..//path> against <non-spec:/p>
  assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
[FAIL] Parsing: <a/..//path> against <non-spec:/p>
  assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
[FAIL] Parsing: <> against <non-spec:/..//p>
  assert_equals: href expected "non-spec:/.//p" but got "non-spec:/..//p"
[FAIL] Parsing: <path> against <non-spec:/..//p>
  assert_equals: href expected "non-spec:/.//path" but got "non-spec:/..//path"
[FAIL] Parsing: <../path> against <non-spec:/.//p>
  assert_equals: href expected "non-spec:/path" but got "non-spec:/./path"
[FAIL] Parsing: <non-special://%E2%80%A0/> without base
  assert_equals: host expected "%E2%80%A0" but got ""
[FAIL] Parsing: <non-special://H%4fSt/path> without base
  assert_equals: host expected "H%4fSt" but got ""
[FAIL] Parsing: <non-special://[1:2:0:0:5:0:0:0]/> without base
  assert_equals: href expected "non-special://[1:2:0:0:5::]/" but got "non-special://[1:2:0:0:5:0:0:0]/"
[FAIL] Parsing: <non-special://[1:2:0:0:0:0:0:3]/> without base
  assert_equals: href expected "non-special://[1:2::3]/" but got "non-special://[1:2:0:0:0:0:0:3]/"
[FAIL] Parsing: <non-special://[1:2::3]:80/> without base
  assert_equals: host expected "[1:2::3]:80" but got ""
[FAIL] Parsing: <non-special://[:80/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <test-a-colon-slash-slash.html> against <a://>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <test-a-colon-slash-slash-b.html> against <a://b>
  Failed to construct 'URL': Invalid URL
[FAIL] Parsing: <non-spec://example.org/test?a#b\0c> without base
  assert_equals: host expected "example.org" but got ""
[FAIL] Parsing: <foo:// !"$%&'()*+,-.;<=>@[\\]^_`{|}~@host/> without base
  assert_equals: href expected "foo://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo:// !\\"$%&'()*+,-.;<=>@[\\\\]^_`{|}~@host/"
[FAIL] Parsing: <wss:// !"$%&'()*+,-.;<=>@[]^_`{|}~@host/> without base
  assert_equals: href expected "wss://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://%20!%22$%&%27()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <foo://joe: !"$%&'()*+,-.:;<=>@[\\]^_`{|}~@host/> without base
  assert_equals: href expected "foo://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo://joe: !\\"$%&'()*+,-.:;<=>@[\\\\]^_`{|}~@host/"
[FAIL] Parsing: <wss://joe: !"$%&'()*+,-.:;<=>@[]^_`{|}~@host/> without base
  assert_equals: href expected "wss://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://joe:%20!%22$%&%27()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
[FAIL] Parsing: <foo://!"$%&'()*+,-.;=_`{}~/> without base
  assert_equals: host expected "!\\"$%&'()*+,-.;=_`{}~" but got ""
[FAIL] Parsing: <wss://!"$&'()*+,-.;=_`{}~/> without base
  assert_equals: href expected "wss://!\\"$&'()*+,-.;=_`{}~/" but got "wss://!\\"$&'()%2A+,-.;=_`{}~/"
[FAIL] Parsing: <foo://host/ !"$%&'()*+,-./:;<=>@[\\]^_`{|}~> without base
  assert_equals: href expected "foo://host/%20!%22$%&'()*+,-./:;%3C=%3E@[\\\\]^_%60%7B|%7D~" but got "foo://host/ !\\"$%&'()*+,-./:;<=>@[\\\\]^_`{|}~"
[FAIL] Parsing: <wss://host/ !"$%&'()*+,-./:;<=>@[\\]^_`{|}~> without base
  assert_equals: href expected "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]^_%60%7B|%7D~" but got "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]%5E_%60%7B%7C%7D~"
[FAIL] Parsing: <foo://host/dir/? !"$%&'()*+,-./:;<=>?@[\\]^_`{|}~> without base
  assert_equals: href expected "foo://host/dir/?%20!%22$%&'()*+,-./:;%3C=%3E?@[\\\\]^_`{|}~" but got "foo://host/dir/?%20!%22$%&%27()*+,-./:;%3C=%3E?@[\\\\]^_`{|}~"
[FAIL] Parsing: <foo://host/dir/# !"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~> without base
  assert_equals: host expected "host" but got ""
[FAIL] Parsing: <abc:rootless> against <abc://host/path>
  assert_equals: href expected "abc:rootless" but got "abc://host/rootless"
[FAIL] Parsing: <abc:rootless> against <abc:/path>
  assert_equals: href expected "abc:rootless" but got "abc:/rootless"
[FAIL] Parsing: <abc:/rooted> against <abc://host/path>
  assert_equals: href expected "abc:/rooted" but got "abc://host/rooted"
[FAIL] Parsing: <https://xn--/> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <data://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <data:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <data://test/a/../b> without base
  assert_equals: href expected "data://test/b" but got "data://test/a/../b"
[FAIL] Parsing: <data://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <data://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <data://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <intent://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <intent:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <intent://test/a/../b> without base
  assert_equals: href expected "intent://test/b" but got "intent://test/a/../b"
[FAIL] Parsing: <intent://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <intent://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <intent://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <urn://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <urn:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <urn://test/a/../b> without base
  assert_equals: href expected "urn://test/b" but got "urn://test/a/../b"
[FAIL] Parsing: <urn://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <urn://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <urn://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <turn://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <turn:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <turn://test/a/../b> without base
  assert_equals: href expected "turn://test/b" but got "turn://test/a/../b"
[FAIL] Parsing: <turn://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <turn://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <turn://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <stun://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <stun:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <stun://test/a/../b> without base
  assert_equals: href expected "stun://test/b" but got "stun://test/a/../b"
[FAIL] Parsing: <stun://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <stun://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <stun://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <w://x:0> without base
  assert_equals: host expected "x:0" but got ""
[FAIL] Parsing: <west://x:0> without base
  assert_equals: host expected "x:0" but got ""
Harness: the test ran to completion.
